api/source/requirements.txt
django
gunicorn
mysqlclient
djangorestframework
python-dotenv
django-cors-headers
